requirements.txt
# Mobile Outlook Agent Requirements

# Core dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Appium and mobile automation
Appium-Python-Client>=3.0.0
selenium>=4.10.0

# LangGraph for agent workflow
langgraph>=0.0.30

# LLM Integrations
langchain-groq>=0.1.0        # Groq LLM via LangChain
anthropic>=0.25.0            # Anthropic Claude API
openai>=1.0.0               # OpenAI GPT API
google-generativeai>=0.3.0  # Google Gemini API

# Database
# sqlite3 is built into Python

# Utilities
python-multipart>=0.0.6
requests>=2.31.0
